Refactor styling. Skjuler paneler som filtreres vekk fremfor å rerendre listen
