--- src/components/_common/expandableProductDetails/expandableProductDetails.module.scss
@@ -1,9 +1,0 @@
-.detailsContainer {
-    margin: 0 0 2rem;
-}
-
-.detailsLoader {
-    justify-content: center;
-    display: flex;
-    margin-bottom: 2rem;
-}
